+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Add stack name to "creating change-set" log message.
+- Add op-map to ex-data when CreateChangeSet fails.
 
 ## v0.26.0 (2025-03-31)